feat(profile): add app version label and logout to Provider profile (MDB-260)
- Add shared ProfileFooter component with version label and logout button
- Version read from Constants.expoConfig (app.config.js), set to 3.0.0
- Reuse ProfileFooter in Client and Provider profile screens for consistency
- Logout: confirmation dialog, auth logout(), router.replace to (auth)
- Provider profile: footer visible in scroll area with same layout as Client

Made-with: Cursor
